+++ AlphaFS.UnitTest/Directory Class/AlphaFS_Directory_CreateSymbolicLink_And_GetLinkTargetInfo.cs
@@ -117,13 +117,13 @@ namespace AlphaFS.UnitTest
 
          using (var rootDir = new TemporaryDirectory(tempPath, MethodBase.GetCurrentMethod().Name))
          {
-            var folderLink = System.IO.Path.Combine(rootDir.Directory.FullName, "FolderLink-ToDestinationFolder");
+            var folderLink = System.IO.Path.Combine(rootDir.Directory.FullName, "FolderLink-ToOriginalFolder");
 
-            var fileInfo = new System.IO.FileInfo(System.IO.Path.Combine(rootDir.Directory.FullName, "DestinationFile.txt"));
+            var fileInfo = new System.IO.FileInfo(System.IO.Path.Combine(rootDir.Directory.FullName, "OriginalFile.txt"));
             using (fileInfo.Create()) {}
 
-            Console.WriteLine("\nInput File Path: [{0}]", fileInfo.FullName);
-            Console.WriteLine("Folder Link    : [{0}]", folderLink);
+            Console.WriteLine("\nInput File Path  : [{0}]", fileInfo.FullName);
+            Console.WriteLine("Input Folder Link: [{0}]", folderLink);
 
 
             var gotException = false;