+++ AlphaFS/Filesystem/Enumerations/MoveOptions.cs
@@ -37,14 +37,14 @@ namespace Alphaleonis.Win32.Filesystem
       /// <para>This value cannot be used if lpNewFileName or lpExistingFileName names a directory.</para>
       /// <para>This value cannot be used if either source or destination names a directory.</para>
       /// </summary>
-      ReplaceExisting = NativeMethods.MOVE_FILE_OPTIONS.MOVE_FILE_REPLACE_EXISTSING,
+      ReplaceExisting = NativeMethods.MOVE_FILE_FLAGS.MOVE_FILE_REPLACE_EXISTSING,
 
 
       /// <summary>MOVE_FILE_COPY_ALLOWED
       /// <para>If the file is to be moved to a different volume, the function simulates the move by using the CopyFile and DeleteFile functions.</para>
       /// <para>This value cannot be used with <see cref="MoveOptions.DelayUntilReboot"/>.</para>
       /// </summary>
-      CopyAllowed = NativeMethods.MOVE_FILE_OPTIONS.MOVE_FILE_COPY_ALLOWED,
+      CopyAllowed = NativeMethods.MOVE_FILE_FLAGS.MOVE_FILE_COPY_ALLOWED,
 
 
       /// <summary>MOVE_FILE_DELAY_UNTIL_REBOOT
@@ -58,7 +58,7 @@ namespace Alphaleonis.Win32.Filesystem
       /// </para>
       /// <para>This value cannot be used with <see cref="MoveOptions.CopyAllowed"/>.</para>
       /// </summary>
-      DelayUntilReboot = NativeMethods.MOVE_FILE_OPTIONS.MOVE_FILE_DELAY_UNTIL_REBOOT,
+      DelayUntilReboot = NativeMethods.MOVE_FILE_FLAGS.MOVE_FILE_DELAY_UNTIL_REBOOT,
 
 
       /// <summary>MOVE_FILE_WRITE_THROUGH
@@ -69,14 +69,14 @@ namespace Alphaleonis.Win32.Filesystem
       /// </para>
       /// <para>This value has no effect if <see cref="MoveOptions.DelayUntilReboot"/> is set.</para>
       /// </summary>
-      WriteThrough = NativeMethods.MOVE_FILE_OPTIONS.MOVE_FILE_WRITE_THROUGH,
+      WriteThrough = NativeMethods.MOVE_FILE_FLAGS.MOVE_FILE_WRITE_THROUGH,
 
 
       /// <summary>MOVE_FILE_CREATE_HARDLINK
       /// <para>Reserved for future use.</para>
       /// </summary>
       [SuppressMessage("Microsoft.Naming", "CA1704:IdentifiersShouldBeSpelledCorrectly", MessageId = "Hardlink")]
-      CreateHardlink = NativeMethods.MOVE_FILE_OPTIONS.MOVE_FILE_CREATE_HARDLINK,
+      CreateHardlink = NativeMethods.MOVE_FILE_FLAGS.MOVE_FILE_CREATE_HARDLINK,
 
 
       /// <summary>MOVE_FILE_FAIL_IF_NOT_TRACKABLE
@@ -84,6 +84,6 @@ namespace Alphaleonis.Win32.Filesystem
       /// <para>This situation can occur if the destination is a volume formatted with the FAT file system.</para>
       /// </summary>
       [SuppressMessage("Microsoft.Naming", "CA1704:IdentifiersShouldBeSpelledCorrectly", MessageId = "Trackable")]
-      FailIfNotTrackable = NativeMethods.MOVE_FILE_OPTIONS.MOVE_FILE_FAIL_IF_NOT_TRACKABLE
+      FailIfNotTrackable = NativeMethods.MOVE_FILE_FLAGS.MOVE_FILE_FAIL_IF_NOT_TRACKABLE
    }
 }