+++ AlphaFS/Filesystem/File Class/File.Open.cs
@@ -995,7 +995,7 @@ namespace Alphaleonis.Win32.Filesystem
 
          try
          {
-            safeHandle = CreateFileCore(transaction, path, attributes, security, mode, rights, share, true, pathFormat);
+            safeHandle = CreateFileCore(transaction, path, attributes, security, mode, rights, share, true, false, pathFormat);
 
             return new FileStream(safeHandle, access, bufferSize ?? NativeMethods.DefaultFileBufferSize, (attributes & ExtendedFileAttributes.Overlapped) != 0);
          }